shifted code from spanconstants to migration
